Test configurators instead of providers
